Initial upload
The last line of this script has some issues that will be fixed shortly. This script is for people new to linux who are managing an apache server. Got tired of doing it myself so I'll have my friend use this script instead of teaching him how to use apache server.
To run, just 
cd /path/to/file 
then
sudo sh Server.sh